fix: internal server error
