+++ product-service/lambda-functions/getProductsById.ts
@@ -1,0 +1,32 @@
+import { APIGatewayProxyHandler } from 'aws-lambda';
+import { products } from './products';
+
+export const handler: APIGatewayProxyHandler = async (event) => {
+  const productId = event.pathParameters?.productId;
+
+  const product = products.find(p => p.id === productId);
+
+  if (!product) {
+    return {
+      statusCode: 404,
+      headers: {
+        "Access-Control-Allow-Origin": "*",
+        "Access-Control-Allow-Methods": "GET",
+        "Access-Control-Allow-Headers": "Content-Type",
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify({ message: 'Product not found' }),
+    };
+  }
+
+  return {
+    statusCode: 200,
+    headers: {
+      "Access-Control-Allow-Origin": "*",
+      "Access-Control-Allow-Methods": "GET",
+      "Access-Control-Allow-Headers": "Content-Type",
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify(product),
+  };
+};
