+++ product-service/lambda-functions/getProductsList.ts
@@ -1,0 +1,15 @@
+import { APIGatewayProxyHandler } from "aws-lambda";
+import { products } from "./products";
+
+export const handler: APIGatewayProxyHandler = async (event) => {
+  return {
+    statusCode: 200,
+    headers: {
+      "Access-Control-Allow-Origin": "*",
+      "Access-Control-Allow-Methods": "GET",
+      "Access-Control-Allow-Headers": "Content-Type",
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify(products),
+  };
+};
