+++ product-service/lambda-functions/products.ts
@@ -1,37 +1,37 @@
 export const products = [
   {
     description: "Short Product Description1",
-    id: "7567ec4b-b10c-48c5-9345-fc73c48a80aa",
+    id: '1',
     price: 24,
     title: "ProductOne",
   },
   {
     description: "Short Product Description7",
-    id: "7567ec4b-b10c-48c5-9345-fc73c48a80a1",
+    id: '2',
     price: 15,
     title: "ProductTitle",
   },
   {
     description: "Short Product Description2",
-    id: "7567ec4b-b10c-48c5-9345-fc73c48a80a3",
+    id: '3',
     price: 23,
     title: "Product",
   },
   {
     description: "Short Product Description4",
-    id: "7567ec4b-b10c-48c5-9345-fc73348a80a1",
+    id: '4',
     price: 15,
     title: "ProductTest",
   },
   {
     description: "Short Product Descriptio1",
-    id: "7567ec4b-b10c-48c5-9445-fc73c48a80a2",
+    id: '5',
     price: 23,
     title: "Product2",
   },
   {
     description: "Short Product Description7",
-    id: "7567ec4b-b10c-45c5-9345-fc73c48a80a1",
+    id: '6',
     price: 15,
     title: "ProductName",
   },
