+++ product-service/package.json
@@ -8,7 +8,8 @@
     "build": "tsc",
     "watch": "tsc -w",
     "test": "jest",
-    "cdk": "cdk"
+    "cdk": "cdk",
+    "cdk:deploy": "npm run build && cdk deploy"
   },
   "devDependencies": {
     "@types/aws-lambda": "^8.10.138",
